+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Added
+- Introduced packaging compilation step that packages a contract to a single package that can be deployed in predictable way.
+
 ## [0.7.1] - 2023-01-04
 
 ## Fixed